Update type of sip_capture.msg field to binary
It seems what the sipcapture module does if this is a text field is
fill the field with the binary contents of the SIP message encoded
into hexidecimal, such that the text of the field contains
e.g. '\x534950'.

To decode this one must say, e.g. decode(substring(msg from 3),'hex')
or convert_from(msg::bytea,'UTF8'::name).  That's pretty gross and
inefficient.

Instead we'll convert this to a binary field that can be converted to
text with convert_from(msg,'UTF8').